+++ zk_ee/src/common_structs/callee_account_properties.rs
